# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Watheq API's Docs",
        "description": "",
        "termsOfService": "",
        "contact": {
            "email": "eng.asgamal@gmail.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/client/login": {
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Check Clients then register or login them",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Request body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ClientLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/ClientLoginResponses"
                        }
                    }
                }
            }
        },
        "/api/auth/client/completeProfile": {
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Complete client profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Logged in User access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Request body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ClientCompleteProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/ClientLoginResponses"
                        }
                    }
                }
            }
        },
        "/api/auth/client/updateProfile": {
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Update client profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Logged in User access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Request body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ClientUpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/ClientLoginResponses"
                        }
                    }
                }
            }
        },
        "/api/auth/client/registerDeviceToken": {
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Register Client device token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Logged in User access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Request body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Device"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/auth/client/logout": {
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Logout client and remove device token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Logged in User access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Request body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RemoveDevice"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ClientCompleteProfileRequest": {
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Ahmad"
                },
                "email": {
                    "type": "string",
                    "example": "ahmad@gmail.com"
                },
                "image": {
                    "type": "string",
                    "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEAYABgAAD//gA7Q1JFQVRPUjogZ2QtanBlZyB2MS4wICh1c2luZyBJSkcgSlBFRyB2ODApLCBxdWFsaXR5ID0gODAK/9sAQwAGBAUGBQQGBgUGBwcGCAoQCgoJCQoUDg8MEBcUGBgXFBYWGh0lHxobIxwWFiAsICMmJykqKRkfLTAtKDAlKCko/9sAQwEHBwcKCAoTCgoTKBoWGigoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgo/8AAEQgAzgFvAwEiAAIRAQMRAf/EAB8AAAEFAQEBAQEBAAAAAAAAAAABAgMEBQYHCAkKC//EALUQAAIBAwMCBAMFBQQEAAABfQECAwAEEQUSITFBBhNRYQcicRQygZGhCCNCscEVUtHwJDNicoIJChYXGBkaJSYnKCkqNDU2Nzg5OkNERUZHSElKU1RVVldYWVpjZGVmZ2hpanN0dXZ3eHl6g4SFhoeIiYqSk5SVlpeYmZqio6Slpqeoqaqys7S1tre4ubrCw8TFxsfIycrS09TV1tfY2drh4uPk5ebn6Onq8fLz9PX29/j5+v/EAB8BAAMBAQEBAQEBAQEAAAAAAAABAgMEBQYHCAkKC//EALURAAIBAgQEAwQHBQQEAAECdwABAgMRBAUhMQYSQVEHYXETIjKBCBRCkaGxwQkjM1LwFWJy0QoWJDThJfEXGBkaJicoKSo1Njc4OTpDREVGR0hJSlNUVVZXWFlaY2RlZmdoaWpzdHV2d3h5eoKDhIWGh4iJipKTlJWWl5iZmqKjpKWmp6ipqrKztLW2t7i5usLDxMXGx8jJytLT1NXW19jZ2uLj5OXm5+jp6vLz9PX29/j5+v/aAAwDAQACEQMRAD8A91J7UZ5pM0VuaWDmmtS000DQuaTmig9aQwzS5pKKYC5opKB0NIBc+tGaSimAuRR0oxzQaAELBVYkgADJJ6D8a838V/GbwnoEz28dzLqt0pwY7EB0B9N5IX8s1wX7SPjG6XUF8N2s7wWccSS3RQ4MrNyEP+yBg49TXz9JfBSBEg45HFJibse86n+0RqMgb+yPD1vEucK1zK0h/JdtcrefGX4gag37q/trND/DBaoB+bAn9a8rbU7g42kKAc8CmNe3bLjzHx044pXJ5j0Cfxv45uTiXxRepkdIpdo+ny4rIudU1+fLXXiW/kJ7G8c5/WuSP2h/75B+vNH2a4JA2Pz7UX8g5jpvO1Jsl9eusjoTcOf61GZ9RdsnXLljjIzct/jXP/YrogEROfoKQ2dzx+6k646Ur+QX8jpUk1aM+ZDrdyjHuLlwT+Rq3Fq/imHHk+I9UGBwFvZP8a4/7PcqfuSD8DUggvF5CzDjPei/kFzu4PGPju2OYvEuqAdh9qY/zNa9n8VviRb4WPW3mH/TaCKT+a15aJr1Bw0o+op6399GP9Y+PcU7hc9ms/jv45t2C3tnpl0B13WpQn8UYfyresP2i7pSBqfhqM+pt7hl/Rga+fxrN6NuWBwMDIqymvScmWBG4wMdqEyrn1JpXx68L3bKt9aapp7Hjc8YkT81Of0r0LQPFeg6+gOkavZ3LH/lksgWQfVDhv0r4dGrwSqPNiKsPQVZg1W23KyExuDwQelMD70PHWivCvgL8QL3UtT/AOEe1S5e7Rome2lkbc6FeSpPUjGcZ6Yr3TNAAaO1BpKYwIoGe9KSKSgAooyKM0AIelIKWikAUtNNLmgAzSUGgUABpBQaTNICQ8mg0dqQ1QCZNFFFIAooooAKDRRQAgGDS0UUAFFFJQAtUNf1a00HRbzVNRcpa2kZkcqMkj0A7knAH1q/XDfG2ynv/hlrMVshd4xHMVAySqOpPHsMn8KGI+W/iN4ml8YeJrnVrqCO0WQKiQx84ReBuOeWxjJrlx9iXG7c3PNQXZJlwTj61v2fgbWLm1E7xLCjDIEuQ35YpN2E97GYLuzj2hbdSAc9BQuqJHxHAuM55r0Dwp8FtX1shpLpLaHqXaMn8uea7u1/Z3tDjz9ZuT67YgM01qKx4OdYftAg44pjatcnOxUAzngV9Q6Z+z34ViCteS6hcY/hMu0H8hWyvwN8DR8jSpvxu5T/AOzVVmUfILahdsflIHHXHNIL++yP3h4r7Fj+CngsuNmmuP8At4k/xqc/CDwdEwxo6tjnBlc/1pWYWR8bHU7zHJHtx0pzavdjlyu7t7V9mxfCrwdGw36DZFTzhgzH9TSt8LvBSyf8i1YN/wABb/GlqFj4w/tm5C4OxvqKVNZbH7yGM85r7Iu/gz4FuBk6DHGT/wA85pF/9mrIvP2fvBdxGfs8N3bk91nLY/A0ncNT5O/tO3kYmW1Q/wC6MU77Tpz4DRMvqRX0RqX7NNgzMdO1eRf9mVD/ADGa4vWf2ftbsixt5xNGO6ru/lz+lLUEmeZRf2Y6qBlee/4U42Fm5GxkYnt0ra1D4YeILIkCFJP904P6gViXnhXWrFd1xYXKJ/e2Ej8xVD5X2Oo+HGsR+CvE8OrRWi3brG8YjeUqBuGCQR3xn25r6f8AAfxB0nxiJIrQSW19Gu97aYgkjuykcMPyPtXxTsu4zg7hivWv2ctN1K+8f2t7GJBZWCSPPJjgbkZQufUk5x7UXCx9WntSUUUwF7UlFFACdDQaU9aT3pAA6UZoooAKTFLRQAUnelpM0AIaSlpKAJPWjtSZwKOtMANFISaXPWkAetA6UE4ooAKKKQDigAOaOwNKelIKAFooooADTZCoQl8bQOd3IxSmqF0/mjOfkB4HrTSEzy65+EHhmTxgmsW0MqWyt5rWJIMJfOQQOoX/AGen06V293YW0pLPDFgHO4gACtMSCKJi3frXBeMNTbypmZiIY1JwDxwKUmooqMeY2pvHHhnRSbX7VJcTr95LaItz6bjgflTIfihpW8FNNvCg77kH6V4bCHurkyYJZzniuq0vRpZEBk+UGuOWJnfQ7I4aFtT2fTvH2g30iI80tozHA+0JhSf94ZA/GutUo8StGwZXGVZTkMPY14Vb6JAB82TXQaGb7Sf+QZdMqnkwv8yN9R0/KtKeIf2iZ4ZbxPUkUh8g8+lIRI0gHQd65a18a2yYTV7WS2f/AJ6RjzI/y+8PyNakHibQZDldVs/+BS7f510KpF9TndOS6Gg6qsvO5vpUzRsU4ULnnrWNc+K9BjJ/4mdu3/XMFyfyFZ998QNKhXZBBe3RHTYgRfzJz+lJ1IrqONKT2R1oQ7fmPPtSAhTjNec3HxC1Bs/YdOgi9PNLOf0xWLe+L/E8+7ZcCDPaGJVx+OCayeJgjZYWb3PYQcNnk0513D7tfPWpah4kvY/n1bUVbr8kzKPyFaHhrxVfS3SabrU8sV2R+7kZyEm+nPDe1EcTGTsKeGlFXPaLmyikHzRIT/tAGqTWWnsSksECn/dBBrDg02WePcsm4Hrk9Kkj0iVThSp+ora5hYq678O/Dmp3UN5NpVnO0ZJKbQA498Yrd0jTrDSrJbfSrK3srYciKCMIM+px1PvWemn3ERyr4PsasxteQNyQ6+ho3Bo1BRmoYJhIMFSrDsamFBOwpNJmiigA60UUUAFJRmg0AGaUU2igB1NNHekoEwzRQaQ0DH0lB6UDvTAU0UUCkwAUZpCaXGKACl96QGgUAFH1oooAKb3pc0mM/SgCOZ8gIpwT1+lZ08hMwC8KKkmm+eRlPX5R7Cq0R3y/N1TirRL3I798jaucAc/WvHfi5qJsLG1tEO2S9nEYH+yOWP8AL869hk+Zz05r50+Mt2b/AOJWk6chOy3RSRn+Jzk/oBWNR6G0HZnSeFNOUQLNIuScFa7K3Q8BRiszTIxHBGqjAAArbthjBrz2egmW4bVmA6D2q9DZsOpGRSWa7yCeB9eTWrFCTnAoVwZmz2BkT5wjZ9az5NBtZcF0APtXRSrtXpUUUbNwB1qmC0MePw/aR4+U/jVxNNtowP3Q9OlascB/ip7xYHNTysdzHezt+QEAqq9rGpPIxWxJFxkd6z7iInOB09KlpjTMy6RE6AVyHi2yhv7B4iQHU7kcHBRuxBrtJxuOO9cl4l02a4hcwp8wBPHeiLsxyWhN8H/iO39o/wDCO+JJV+3D5ba5Y4FwP7p/2v517jHskTcuOa+C/FbzWurYctHKhyCCQVIPXNfSHwJ+Ji+I7H+ytXkA1i2XG4nmeMcB/r6/nXpUp3R5lRcsj2NowMjGDULRgAnHNWAcpkHP601udoraxCZnOpQ7gPmq0pDAEdDTpo8qc9qgtTlCp6g1EkDJqKDRmgQdKTNFIf0oAWj1pO1JQAGilpKACg9KKSgANAopKAHg+tFJRTAXNGaQUUgF60UlFACjig0lGKAFFGaTFLigAprttRj6AmnVFdHFtL/umgTMGRmEY5571NEu2MNzk9arHMhCjrnFWHO1cHp2qgRG2NrMTjFfLurz/wBp/Hadslgl1s/BEx/SvpjU7lLWxkmlIVFQkknGBXyz4Xgum8ayaxIsTySvNMkJlAdg2cHHYc9axqbFx3R7Ld6pZ6Taia+mWMHhQerfQVkH4jaOkwiDTH0ZE3LXMX1jeapdC51VYZkU/LCZGVVH4Cntb6dKvlSWOnZ9FLFv0Fcei3O5s9B0b4h6NcMqRXSh/R+DXXWHiKKbGxwc+hrxOPwxZxMLiztZRIASAI1PPt8wP9asaP4gaxvUtpYjE+dpBOAPcg0PTYadz3k3yyx9cn2qpe6isKDtjtisLw7qUV9GmySNiegDjP5UzxJqFnawNmRZJOixxkMzMeAAPr3PAqbl6Et54t+zgkKdqnBPvXI6j8WZopPLt9KuZJCcAHHP4cmsrUNXuWJWK2gQk8BpDIR+WB+tMtru0sykl7cWsMxGWLSjA+g2/wBaak+gpWR0Vn461q75/sK6RSMg7Dj9atW/jO9jbGp6ZNHCeMheR7+mK5+DXZJkJ0/XdJuWXonlHP8A6FUsWt30kwhu2sUc8ASQMAfxDkUOXcUVfVHZQajbaggltJAT3XPP5VKQWU/T8q5NpreyQXWoW8MGHXM1sTtAPVmzggCtpdZs5IpGs5kuQhKs0bZAI7Z/EVHI3sXzpbnj3xy0wQXFrdouN5Ksa840rU7nTL+2v7KUxXMDB0cdiPX2POa9k+JZfVNL3ywE26N1Hr/vH+leFz5EzqowoJAFdVL3UcGI3ufcXws8c23i3w/BdwkLMPknhzzG/cfTuD6V3OA6h1PXtXxJ8FfENx4c8RvcqZHtZFCTwqfvLnr9R2+tfZGgarb6pp8NzZyCSKRdyMOMiumE1LQz5WopsvnjIqlH+7uCP4WHFaLqccjBxVO6iI2yD+A1bDoPzmjPNIpGBS1JItFJSZoAWkooJoAM0lGaO/FAB0FAopM0ABoFBpKAHmgUUlMBaBRRSAKKKKAClpKWgAzxRSUtABUN3zby/wC4alPtTJQWjdf7ykUhM52IEMSByvNLdXIjwDyfSiLKqQDyaihhBlLvziqYI5H4kW8+oeEdViaQRb4G25OAMc5P5V4b4EhceI5r1huRLWMZ653BRn9DXu3xNmUeHLiPIXzV8vJP97ivJvhLELzRblpkwYyIY3HUrktg/QsfzrnrOyN6WrL/AIheSSJ/J3gY42jJrmdG0+e9ungvpJNPtDwoRsFye7tXpo0wiMqE3Z75qAabcRsdqIV/2mUVxKVjs5bo8fbRNZs7hoobXz1Dsvnld+SVK5DencehrfutM+wWML3l01+RFmZGOWiY8DDdx0zmu6lMsKlf3SAdg+f5VUWyuJpE3IDGevmLww/3e/41p7W5KpJbF74V+GLBtDtr25tv9JkyVcMysq9B0qTx34bs9HgXUNMtxbyggllXPccnPbGc59q7fwzCsdnHGgAVVAA9Kt6xAk6rE43Cs76myjoeViykvrGIpHLayOT5juhOwA44HqffpWdr3hfSrnw+Le1WYakknmF5kO6Xgg5bHvkdq9Fn064+dR+8B5+X5S319/esa4sHikz5kkPPSRNwpxlZ3FKDasedaf4J1S5vEfWZvtKxytIIw5JO45PPbJ5rY07SdV0e9xLcCaybOIpgXKj2NdraW8nAFxET/uEH+dakemrKAGIZgc5waU5OQQgoLQ5LUIbu+0e8iEKiExNhQOpxxW14d8J22g2gt3mkuvm3fPgLnAH3f+A966iKxSFVwOV5wKZcocZqVJx0RTipO7OY8aWi3WgXcYAwEyAK+fYNCOoaq8cLpg4Y7jgYPevpPVU8yzlTHBXFeFyWRttfKxAhd5VGH3lIzj69CMHrWtOW5E4JtXJ9C0U6fqDRr8yFgA4BAYe1ew/DfWbjQbx7S4O6yc7gP7ue4/qK4rTtLuhO/wBofc6fNGyqdrg9CM9PT8K6VYGis43YFpIXBBxQpuLui3BSjy2PoC0uI7qFHRgyMAQRyCKfIuQRgcjpXknwg12aKXXdNnl8yK1vyYQ3/LOORQ4X6ZLfnXr0ZEibs8EcEV6MXzK55bVnYoKNp2njFLT7lGHzcZFMByM02SxaSg9aDSEBNJQO9FABSZoNJigBc0UUh60ALRSGigQ+kFFGaYwpaSjikwFI5paTrR0oAM0UDmigBaQ0lGaAFHWkPWlHSk780gMPYEaVf7pIzUTbY4MjvUr/ADLkZy7EmqmpyBYyqduKpiR5j8Wr0f2dGjfdklC9fQEj9cVi/Cyy+yeFoAw2u8rsR+OP6VL8VIZdQm0+yt2Cu8uQx6DAJyfyrU8KRfZ9FtUYY3KZAPrzXJXOukjp4Y0ZQSMn6U97GN/+WYyfam2mOMVsWkYkA54rksda0Mj+yreEbhGo+orKu9okOAMmuu1Ro4LdtxAGK5GANPeM7fcP3Riny6ldDpNBTZbDPYVYvyVkDY+WptNX92BjgjNO1CI+XznBptDiQ2iLK2V5qeWyWXAZVP1FZMTyWzeYmcDt61rafqazDEnU9KleY5XQz+zkGdqKPwqzBbqnGP0rQTY+ORzSTIFwafLYi5RmAAyKy7nBznpWhcNgErWXO+7cKhlIy7zbtYHGCK80GkfaPEwQx5hYO5GM9Gr0a9kCKxPQdaxdKQSagMgZBOD3xTjoTLc1NO0g3UjxQRSRxKCrSOeCc9FBrSNhELJbNjlUO6SQ+1b2lny7ZVkzxXFfFvUv7H8J6nJC2yW4QRIQecscH8lzVJXByscb8Fdeju/H/ieHdlLpjNF6FUYqP0Ir6F0vUhAyxSnMZ4HtXzn8EfDJtok1uSTE8gLRIP7p4/oa98Ta0K5wVIHIr0obWPKeruda0Yki3Agg81R2lGKnp2qppt40JEbnKHofSta4jDxAjrWgFSikz60ZqRATRRSGgBaKbmlBoAWkpKOKAFzmkNGcUlICSjNJRVAAOaWkHFFIBaOaKKACigUdqACikzRmgBelMmbETnHIU06oL19lu3qeKBMzBhW/3RgVlX7FhJ6jmtN+I/c1janJshkYd6bBHnfiQLd6gilc8MCfQbSD/OtcrFEI/KxsCAAD0qldxeZese4jbmpJiIZIhklCu3OOlctSN9TenKzsbWntzz0rft5dkZPU1yumy5wD0rTacRQli2F7kmuRs7k9CDV5nuZCCfkX9azzfWttPGGniV2+UKzgHP0plzqMJY88evQGuW1t7Z5pJ2ihleOPcXZQwQU0NyVj2DS7mNIw2Mk8A1W1/XbS3jaS5lSCIcbnbAzXm2j+JrxYUjW1kkTaCrngYP1q6urXN/dJi1Bk5A3DO0DrVMFNLc6WHXdOvoVSzl81m4GEZf5gVdjsiLZWQ4cVnafcb5IzKBjaPmP9K2TfRRptzwPzqbD50yTTr0qNrnBBrQluNwBz2Nc7LdQh1kU4VjjpV/zN0eM5FTfuG465lzkbuTWfI3BJpLuTZ8xGQOprPkud5baflA4Prmkx3sZ+q3AUEMOtUbO5hsna7uZljhhUs7seFAp15KNziUDKjua4P4mXpg8JPHHlDcSKpHqo5P6gVcI30MpzsrnqjfFLwmkJLavbHA/hVmP8q8a+IPi1viBr1rp+lCWOwjJCFhy7Hq5HYD09K8sDM2EHc13mi3ljolqrrGy3CRfO+Pn3np+HFdMKajqcnt5TVke8+HrNNJ0+1toMiKFVjUkYPArtNGkEitC5BI+ZSK8L+G3jqbVp003VVTzmVnikUnJx1BB/OvXNLnYMjxfeX07kVvHyMmdPzG+OMVr6ddEERycqehNVI4Rd2iXEXcdPSooc8oc5HIrQk2bqMZ3r+QquDkVPp8yzIUkPzj9ajuIzG/8AsmgCM0CkzxSE5pALmjPtTc0fjQA7NGaaeKM0AOpM0mfek/GgCWgUnNHNMB1GaTPFJnJFIB1H8qKKACj86KSgBaKQe9FACg1Q1GTc4jHbrV122qSe1ZbfMS7ck800JladsKcelYOsHEOPWtyciuf1nLED16UmETnXhLSuV5JG0fmKta9ZPBozO64KYfjqeen5ZrX0fTvOugXHyqMn69qm8Ur5tlNFgYKkVHLoVezOO0yfc2AwyRkD/Co9TuJpLgxKxESDPI71naRK3yEr04YenrWpeAb4TGADnAPfOK8+aszuhK6MdUadmVj8vUZ6sff0+lXLLTrXy8Xsitk7tme9U9V0FJY2uYZ7iGX1jkI7+nTmpNH0E3RxDqE6yDkiRQ4/pQtTaELnUWFhY7U+yzL8v3Y2fjNWobGytZZHlkQ7yCQDk9c1lnw/eoeJo3AxgbSp/rVmDRb5+ksaL/ulj/SnZmvsUaB+xhy0UqqxO7JOMnFV7mQpkpNHsxjL45/EdDVe40WOMYvLmWRj2X5R+lZsXh61vLgK8O6HvvYtkenJpN2E6aRciuoriV40uI3xyQmO/wD9etOwucJ5GcsP1qtPokNtEDZxpE68qEUAVNpi4jlZiCWOQPQ96l6maVmF28m4bThs8A96y9xZCAnQY5+v8qt3rqJBuYgdCe3SsuaXaV805K/xDgdf/wBVEQkyleEFmcAbhxz04615r8V5ma1soehdmYKD2GK7+5kEkhEbjK5J44yTxXmnxVLR6raQjqlvvIx0yxraC95GFTWLPP1j+QsTjHTPephez4VS+VXt61NOFaMIh4znpyKk0jTxPfRLMCU3cqCATXTY4OWSehtaA97AHv7cMgQCC3bphnIBI/DNfTuhotvbWwVspsXJ/CvHPD1nH4g1y2s4lWPSbD95IVxhn6Bcjqele8aVpoEZYIAD0UdvarSNHodDoN19nkKE5jate/sSUFzbEMw+bA71zEAa0uEZh8nvXV2czIq7jmJvun0rVdiDP2n5ZUztPINaSEXNvg43inz2qrkoP3T8kf3TVRVNvKD/AA9jTsFyJhgkGkq1doHAlUY7EVTz6VDAdnFJmkJ4pM0DHZpKaTRmgBSaM00NmjNAifNGaaGzSigB1JnmgUnegB46UUlKKAFpCKWkzQAtBopCTQBXvWwij+8azp5No460/UJ1MxA6LxWc0nmPgdBTuBJgnk1mNbG81AgD5EHJrUEbuNqgknsKu21sttFjOXPLEetFriKyW4t7f07msrVUMqfKM9zWxcZOeapzJ8h784zQxI8iuYGstcuIiD5THchHqeuffp+tPuLpjcQjOAOMhufSuo1nT1muiQMPngmuHuALa8kMpXanI6Vx1oWdzrpS6HULKskIUdhjBqjJIdOlE2SFJwSKp6fdl3SMjJI4bpx2rq7aBXUBwDn1rl+FnXF32M7/AITe0iXZcXOwj1U/SprbxlZz5SG53be2CKfe+F7K7aQiNCz9TirekeFLOzTy44EwOmef1q+YvmY23uftz5jDMucZxxW1bxLANvBPrU8NksK/KAM/hVXUSsa4YgY6jNRuDYT3aKGGVJGf/wBVc+t40W4kdDn65/8ArUly0pcs2djcYPb3qldutuFRIzmTHU0WM+Yt30+4OiEsxw3I+7WVdkMNgBG4E5IwRjnH1p7tLKWOf3oXOB9SP6VGsJuJ/LRucDIA6A/1q4xJbuM0W2N7eK7r+7Vcggde9eYfGSC5tPGR+0rgTW6NF6bOf6g19GeD9CWV9xQCNcGQjoT6D61yH7SHhRtR0KDVrWMtcWD/ADgDrG3B/I4P51006f2jCrPSyPmsKcCQfxYzXReFdMbULyKN41YHqWHHT2qobJkigt9jCXoAFyWP+T+lev8Aw98LtYafHJdJtnZckH+H/IFaGV7HUeFtNhsLeMRRIm1BgAcA/wCTXqegIPKjDAYxXD2SYKoeg45rtNGk8uNADkdD7VpAzkaGoWBkDFRuXvjtTdGuPKc2lx0PKH1FatvJ8h77qg1DTBInmQZDjnjqD61pbqQi/HlMI5yp6GmzwA/Lxg9Ce1Q6fc+bGYZxiZeo9fcVcUbRhhlTQMqCPYpVuR0rKmXy5Cp7VsyHyye4Pes3UFwQ4GRSY0VC3NIWpmTmkzUXGO3Zo3UzdSZouBJmlBqLPpS5piLQp1NpRQA4UKRk03PFC+1AEg6UU0GlBoAXtRR1paAAVXvZxbwM2fmPA+tT1m6nA9zMiKcIoyTQBiSyNI2OTmr9lZSFQzjYp9e9X7ayhtgDGuX7s3WrOe5HNUok3I1RIkwoqJzmpJXGehquxPXBApiI5BkH1qo44Iq0VJGQfwqpNhcs7YAGST2pMDE1SD5i2QBivP8AX9CvLiG7vra3YwAAkq3ORxkA9e1er2mltqMiyz5W2ByqHq/ufQe1bL2UUahSoIPAz0pOHMtS1JrY+b9FuV85SsoOMDce9d7p18km1EbcB1JORnqRWV8RfCUmlNLqWjw5tHB8xVHMJ7ke38qwPD2r7SElZg6gYV+/b/P1rhq03FnZSqJnrFmVGGOMYzV+NogM7h0z1rkLHU/NAUnqvpU4virddyZGBjrWNjpUrnR3V3HGgIbmsO9uknTJzk9D1/Kql/fhEVnJwW7HoMVzuoan5Rf7PuYHGFHqe+KdrilJFy9ukUbjkqSVI/P8qoCNpZ13swWQDYRnC8f1zVeJGuXkYuwVSCX5O5vb1rZsbaXUpktLWCedyeAiksfUn+6vucVagYORDhptiQ4Qv37n6Y711/hTwy96/mKuyHPzykYH0Hqa6Dw54KEEUb6qynABECHIHsx/oPzrt4o1jRUjUKgGFVRgCuiFBvWRnKrbRFS0s4rO2SG3QLGv6+9V9QskureSGZQyOpUgjgg1qFaYy8811qKWxzydzkbDw9ZQkbrWMFeAwUcVZv8Aw1DcJvtyEkHfsfrXR/ZWXnBpyoV9qdkybnmc9lPZXIjuIyp6A9j9DWxpk2wDn8a7C5tYbmMpNGrrWDcaK9uxa0bcv91utRy2YN3NKymyBjpWtE4Yjk1y1tM1tLtkVlz2I6V0FtJuUEf/AK6pMLFu4gSZ1dCEdOQ3+NOjZm/dyDaR+VMRjkE8VNwR70BsMlj+Vg3WqFwhMZQ5/GtLOeGGRVadc89TUspM5xuCQeoppORU16oSdsdDzVbNZsB1JnmmlqaTSAkz6UuaiBNBbrTA0M0oNMzS1QhxPFKtMBpwNAD6Xr0ptKDzQAoFOzTahuJdi4HU/wAqAFmm25CcmmoCVBPeokXJG4f/AF6twoAADx7VaRLY0AgUYGDxVjYCOnPejZg5647VQiqUGBnj+tI6KU54A5HvVgx47ZIqGYhR9OeaQFS4ChSelU7S2F5IJZRm3B+RT/GfU+woydRvPs8ZPkJzKw7+wraKBV2qAMDoOwpJXHsNVznAxtHtSNggseg6U4LyAKZMcg+g4FWIzb//AI87olQ3yMSNu7PHp3+leTXfgpb2Nr7w+0ayr8s9uMgKxGcoW9s/L09DXsIHJqpbR2dvdy29vH5UjfOwVcLzn8OprKcObctPl1PCNI1ZBII0nUuhKsrghlYcEEexzWgL9toEbZUn7w+tY2saWNM8f67pksYCSzfa4c91fk4/HP5V6d8PPA1hrWmPdXiF3jmMeD9AR/OuKMLy5Trk3GKZw7X1xcuIoLeR3bG3AztNbmh+BNevrhJp7ZI06kzt5a5P0BY/kK9jt9Fs9I8uKztUEhHy7QK0rQtJvDoEkQ4IFdMaKW5k6rexw2lfDSxjZZNZvJb9hyIo18mIfgCWP4mu30+wtbCAQWNtFbwj+CNAB+PrVsLT1WtlFIzbbEVaft4pwxmnVRJERUcilmVFyGLAVM55wvLfyp9tH87Oe3A+vehgSAfIQRnHemyxDHA4qYphOKaoymKLksq7MHgUphDdRg1N91sHp61Kwx1ouIybi1DDEqBx6kVWitjA2YWOzuh5/XtW2y5qrLD3HBoKRAsh6MMVKrkH2qI7h94cUA46H/CgZcVgRS7QRxVVZcEdvftVmNsnJqRWKV5YQzjLDa/ZhWDd20lsTu5Xsw711xXcABgjvVO6iV1IYAg9qloZyeeaCeOKmv7U20nGTGehqqTxUDH596PxpgPFGaQjTHWlFN6UoqwHZpelNpRQAqNng0+mCnKc0AOJqtjzCX9elSXDFYjjvxToEwEU88CqiSxIU3DnOB+pq4kY4I7d6ZEMGrCjPTgGrENIwRntxQFJ7VYSEHrzxRIdq4oAqzbY0LkgAdc1zd/dy3cnk2ueTjPr9KvapI9xJ5edqenrV2xsEtYweDIw+9S3dhpkOl2K2VvtGC7csfU1ZYY4qy6AEAd6a0QAwKewXK4GBn1qvP0UD1q3IPmAqK4iG3r0oEU3HGR/CcH8ayr/AEs3F9BdwTvDIhyy9nx0/wAcdP51t4BAPYnaRUTpskwTnnaf8aTVxnl/xj0F3msfEtooL2CmK5UdWhJ6/hkn8a734NSK2i3wRsr5qsPoV/8ArVo6jYR3WmzxyANFIhR0YcEY5rJ+DekSaFperabLP55tbzyFk/vKEVlPscMPyrL2dqikjXnvTsd1eWq3DK27aw9qWKFYlwv/AOulc4lAPTFInDNjp6VsZkgFL3FQT3AiOMEt+lTRxOwDSvwedq8frQ2MXIzjv6CngMTzwPbrTlUDhQFHtUoXauTzSuBEExwOCen+NWEUKAB0FIg6k9TTxSJYj0xKeetNUYNAkMlA2H1pQ26MGlboRTYf9UR6UwAGmtg9aX1qF2wSDQhoR0DDpVWSFhkrV9VyM0FQBTHcygSDzwacshQ+1XZYVcfMKqSReWcZyKBItQTrIOoJ9KklUEZ9qzTGR86NtNWLS6MjGKQZIHJ9aTKK13CrqY5PunoTXN3ELQSsj9R+tdnPEJFPY9qwtTgE0JPAkTv61EkO5iZozTM806swP//Z"
                }
            },
            "type": "object",
            "xml": {
                "name": "ClientCompleteProfileRequest"
            }
        },
        "ClientLoginRequest": {
            "properties": {
                "phone": {
                    "type": "string",
                    "example": "0532579547"
                }
            },
            "type": "object",
            "xml": {
                "name": "ClientLoginRequest"
            }
        },
        "ClientLoginResponses": {
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "name": {
                    "type": "string",
                    "example": "Ahmad"
                },
                "email": {
                    "type": "string",
                    "example": "ahmad@gmail.com"
                },
                "image": {
                    "type": "string",
                    "example": "uploads/ahmad.jpg"
                },
                "phone": {
                    "type": "string",
                    "example": "0532579547"
                },
                "created_at": {
                    "type": "integer",
                    "example": 1511649843
                },
                "isCompleteProfile": {
                    "type": "boolean",
                    "example": 1
                },
                "token": {
                    "type": "string",
                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOjIsImlzcyI6Imh0dHA6Ly8xMjcuMC4wLjE6ODAwMC9hcGkvY2xpZW50L2xvZ2luIiwiaWF0IjoxNTExNzI2MTkxLCJleHAiOjE1MTE3Mjk3OTEsIm5iZiI6MTUxMTcyNjE5MSwianRpIjoiSXFMck1NTU5sTlNFSVdBayJ9.B765ouFHZLcA1KyRVoDghO3zlvGey8glZ7wxIvL89To"
                }
            },
            "type": "object",
            "xml": {
                "name": "ClientLoginResponses"
            }
        },
...
    },
...
}